chore: split file in model
